requirements.txt
fastapi
uvicorn[standard]
jinja2
line-bot-sdk
openai==0.28